refactor: improved file structure on the wallet routes
